Move SampleProcessor into separate module
