+++ common.py
@@ -1,0 +1,7 @@
+import numpy as np
+import numpy.typing as npt
+
+SamplesT = npt.NDArray[np.complex128]
+"""Alias for sample arrays"""
+
+FloatArray = npt.NDArray[np.float64]
